git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@10167 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_marineview/MarineGUI.h
@@ -75,8 +75,8 @@ protected:
   inline void cb_ToggleHash_i();
   static void cb_ToggleHash(Fl_Widget*);
 
-  inline void cb_HashDelta_i(int);
-  static void cb_HashDelta(Fl_Widget*, int);
+  inline void cb_HashDelta_i(unsigned long int);
+  static void cb_HashDelta(Fl_Widget*, unsigned long int);
 
   inline void cb_HashShade_i(int);
   static void cb_HashShade(Fl_Widget*, int);
